requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
google-api-python-client>=2.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
httpx>=0.24.0
